gtali: Added GtkApplication (and application menu) support
This is an attempt at fixing bug
https://bugzilla.gnome.org/show_bug.cgi?id=666819

Adding a GtkApplication object is done quite trivially, and only
required adding a static variable (I guess it would be possible to do
without that by modifying existing functions parametres).

The problematic thing is going from GtkMenu to GMenu. What I didn't
manage to solve here is how to make the "Undo" item non-sensitive: I
don't even know if it's possible to do that in an application
menu. While I can imagine some workaround (e.g. removing the item from
the menu when it should not be sensitive) , I wonder if it really makes
sense to put Undo in this menu at all. Additionally, the difference between
GActionEntry and GtkActionEntry forced me to keep the old
"action_entry" structure only for the "Roll" accelerator.