Move scores from /var/games to home directory
The score system was originally designed for multi-user systems where
storing scores in a central location allowed different users on the same
computer at different times to compete amongst each other for high
scores. This might still be fun in a computer lab, but nowadays most
users have individual machines, and we are more concerned about ensuring
a user's scores are his own data, on his home partition, to remain even
after blowing away / for an OS upgrade.

All of the modernized games have already switched to scores in the home
directory: time for the rest to do so as well.

Also, I get to delete a lot of code, and that's fun.

https://bugzilla.gnome.org/show_bug.cgi?id=690430

https://bugzilla.gnome.org/show_bug.cgi?id=690432