help: Drop mentions of need to restart game for game type
It had grammar issues, but looking at it, it doesn't feel like something
we need to spell out here under this section, as this really applies to
changing anything in the preferences, not just game type. So it should
be added as a generic note outside any specific preferences section.
As we don't have that right now, and the game itself tells about it to
the user after closing the preferences with something changed, simply
drop that sentence fully instead.

Closes: https://gitlab.gnome.org/GNOME/tali/issues/14